Develop (#170)
* a733L renames

* style

* comments

* A733: added PPU base address

* A733: add PPU registwers

* Add ppu pit definitions - https://github.com/ARM-software/SCP-firmware/blob/master/module/ppu_v1/src/ppu_v1.h

* add cpuidle added

* GUI: разделение на user и framework части

* Добавлен параметр WITHLOCKFULL -

* custom config update: spell checking

* genregdef: add class comments support
a733: add r_cpucfg stub

* a733: work in progress

* save state

* a733: save state

* a733 smp: save state

* GUI: save state

* GUI: save state

* раздельные алрибуты для буферов приёма и передачи по DMA

* a733: Update PPU & R_CPUCFG registers

* A733: add r_cpucfg registers

* custom config update

* a733: save code state

* a733: в режиме SMP питание ядер включается

* save sources state

* style

* GUI: update submodule

* build fix

* A733: update ppu

* update wait functions prottypes

* a733: update PPU interface

* style

* code formatting

* fix gec cpu id for A733

* custom config update

* style

* add openocd config

* a733: aarch64 smp startup: style

* style

* style, renames

* style, убрал лишние поля периферии

* style, убрал опечатки

* a733: запуск и в aarch32  и в aarch64 (core0..core5)

* style, comments

* a733 smp: style

* a733 smp: code state

* style

* a733: запускается core6 & core7

* aarch64 init update

* style

* style

* FEL paiload update

* A733: Удаление отладочного и лишнего

* custom config update:

* aarch64: fix exception stack frame layout (and size)

* Add SVC handler prototype

* закомментировал тест

* build fix

* подготовка к работе с обработчиком SVC в aarch32

* mistype

* a733: back up cpu pll init

* style

* renames

* style. comments

* a733: update xfel bootloaders (1200 MHz cpus clock)

* style, comments

* a733: add DSU P:: interface

* style

* Add coremark test

* save coremark config

* fix build

* A733 CPU PLL init style

* pre-built images update

* tests update

* a733 clocks: style

* cpu burning test (all cores coremark)

* Linux & AMS1117: SWR support

* mistypes

* добавил svc_handler в интерфейс переклчателя задач

* task_ticker interface added

* scheduler: обработчик SVC

* пробная реализация delay

* save state

* scheduler: передача параметров через SVC

* scheduler: less PRINTFs

* scheduler test

* style

* scheduler: работвет

* scheduler: работа в aarch32

* scheduler: работа в aarch32

* gui save state

* test software update

* scheduler off

* fix linux build

* add event interface prototypes

* Замена local_delay_ms на ..us там где перепланировщика нет или нежелателен

* style

* style

* comments

* fix risc-v build

* submodule update

* Fix __CORTEX_A symbol usage

* style

* comments

* style

* style

* style

* bootloader build fix

* t113: update xfel payload

* style

* style

* добавлена кастомная конфигурация: A733 + Linux

* Update product_template.h

* style, одготовка к ожиданию регистьров в scheduler

* style

* save state

* save state

* save state

* save state

* save state

* восстановление работы wait32/wait8

* отказ от debugtimer как признака таймаута

* debug cycles counter обрезан до 32 бит

* удаление неиспользуемыз функций доступа к регистрам процессора Cortex-A55

* Add stubs for Cortex-M

* renames

* renames

* add risc-v context switch stubs

* custom config update

* renames, исправление расчета таймаутов

* fix time limits

* renames

* style

* переименования в aarch64

* style, fix risc-v build

* renames, style

* comments

* Update risc-v nuild

* fix formats.c in 32-bit mode

* tests update

* fix wait8 fsyscalls

* fix non-rtos waitX implementation

* style

* Добавил функцию ожидания непустого списка в системах с планировщиком

* style

* scheduler timeout handle

* sae stte

* save state

* save state

* save state

* renames

* test

* Возможность обработки звука вынесена в отдельный поток

* bulld fix: не все файлы закоммичены

* scheduler: add lock support im local_waitlist

* Fix non-smp systems build

* build fix

* comments

* save sources state

* comments

* comments

* scheduler: more debug, удалил events

* comments

* more debug: add thread name

* инициализация пользовательских потоков перенесена в места после инициализации планировщика

* scheduler: more debugs

* save scheduler with debugs

* scheduler: more debug

* cou support functions

* mistype fix

* style

* comments

* build fix

* style

* style

* mistype

* code style

* удаление неиспользуемого прототипа

* comments

* comments

* comments

* submodules update

* code style

* Подготовка к работе с YUV видеобуфером

* Подготовка к работе с YUV видеобуфером

* тесты выдачи uv (пока не всё)

* yuv tests update

* tests update

* build fix

* custom config update

* Добавил управление magnetic loop в меню и передача у тюнер

* добавил отображение состояния MLA

* Обработка состояний от MLA

* custom config update. брал баг с большим шагом перестройки в NFM (оа прикрыл)

* build fix

* style

* style, renames

* Устранил зависание в передаче управления на MLA при частой смене частоты

* Отображение состояния MLA

* custom config update

* fix local_delay_ms

* encoder: style, tests update

* encoder: устранил ускоренный шаг при работе в режиме снижения чувствителдьности валкодера

* encoder: style, коррекция таблицы скоростей

* 1024*600 display layout update

* Восстановление работоспособности теста LVGL

* scheuler: дополнительная отдаджочная печать. приём есть.

* style

* fix scheduler usage

* comments

* Удаление веток с использованием FreeRTOS

* scheduler style

* save state

* style

* save state

* Удаление IRQL_IDLE

* style

* удалил неиспользующиеся функции и константы

* a733: more debug

* gic: more debug

* gic enable

* fix a733 build

* save state

* encoder: style (подготовка к улучшению сглаживания расчета скорости

* encoder: style

* bootloaer build fix

* custom config update

* Попытка исправить работу с SPI на velociraptor

* custom config update (ANU8822 volume control)

* pre-built image update

* custom config update, build fix

* costom config update

* comments, code style

* comments

* style

* pre-built image update

* build fix

* build fix for non-dsp builds

* Fix Cortex-M7 statt-up

* test loops count update

* custom config update (usb host test)

* custom config update, comments

* Allwinner V3s init update

* Allwinner V3s init update

* t507: added CPUIDLE fields. Test added

---------

+++ src/usb/usbd_dfu.c
@@ -159,10 +159,10 @@ typedef struct
   const char* pStrDesc;
   USBD_StatusTypeDef (* Init)     (void);
   USBD_StatusTypeDef (* DeInit)   (void);
-  USBD_StatusTypeDef (* Erase)    (uint32_t Add);
-  USBD_StatusTypeDef (* Write)    (uint8_t *src, uint32_t dest, uint32_t Len);
-  uint8_t* (* Read)     (uint32_t src, uint8_t *dest, uint32_t Len);
-  USBD_StatusTypeDef (* GetStatus)(uint32_t Add, uint8_t cmd, uint8_t *buff);
+  USBD_StatusTypeDef (* Erase)    (uintptr_t Add);
+  USBD_StatusTypeDef (* Write)    (uint8_t *src, uintptr_t dest, uint32_t Len);
+  uint8_t* (* Read)     (uintptr_t src, uint8_t *dest, uint32_t Len);
+  USBD_StatusTypeDef (* GetStatus)(uintptr_t Add, uint8_t cmd, uint8_t *buff);
 }
 USBD_DFU_MediaTypeDef;
 /**
@@ -202,7 +202,7 @@ static USBD_StatusTypeDef MEM_If_DeInit_HS(void)
   * @retval USBD_OK if operation is successful, MAL_FAIL else.
   */
 // called on each PAGESIZE region (see strFlashDesc)
-static USBD_StatusTypeDef MEM_If_Erase_HS(uint32_t Addr)
+static USBD_StatusTypeDef MEM_If_Erase_HS(uintptr_t Addr)
 {
 	//PRINTF(PSTR("MEM_If_Erase_HS: addr=%08lX\n"), Addr);
 	if (0)
@@ -238,7 +238,7 @@ static USBD_StatusTypeDef MEM_If_Erase_HS(uint32_t Addr)
   * @param  Len: Number of data to be written (in bytes).
   * @retval USBD_OK if operation is successful, MAL_FAIL else.
   */
-static USBD_StatusTypeDef MEM_If_Write_HS(uint8_t *src, uint32_t dest, uint32_t Len)
+static USBD_StatusTypeDef MEM_If_Write_HS(uint8_t *src, uintptr_t dest, uint32_t Len)
 {
 	//PRINTF(PSTR("MEM_If_Write_HS: addr=%08lX, len=%03lX\n"), dest, Len);
 	if (0)
@@ -294,7 +294,7 @@ static USBD_StatusTypeDef MEM_If_Write_HS(uint8_t *src, uint32_t dest, uint32_t 
   * @param  Len: Number of data to be read (in bytes).
   * @retval Pointer to the physical address where data should be read.
   */
-static uint8_t *MEM_If_Read_HS(uint32_t src, uint8_t *dest, uint32_t Len)
+static uint8_t *MEM_If_Read_HS(uintptr_t src, uint8_t *dest, uint32_t Len)
 {
 #if WIHSPIDFHW || WIHSPIDFSW
 	//PRINTF(PSTR("MEM_If_Read_HS: src=%08lX, dest=%p, len=%08lX\n"), src, dest, Len);
@@ -317,7 +317,7 @@ static uint8_t *MEM_If_Read_HS(uint32_t src, uint8_t *dest, uint32_t Len)
   */
 //#define DFU_MEDIA_ERASE                0x00
 //#define DFU_MEDIA_PROGRAM              0x01
-static USBD_StatusTypeDef MEM_If_GetStatus_HS(uint32_t Addr, uint8_t Cmd, uint8_t *buffer)
+static USBD_StatusTypeDef MEM_If_GetStatus_HS(uintptr_t Addr, uint8_t Cmd, uint8_t *buffer)
 {
 	//PRINTF(PSTR("MEM_If_GetStatus_HS: Addr=%08lX\n"), Addr);
 	uint_fast8_t st;
